Corrections form + requete ajax

+++ asset/php/suppr_besoin.php
@@ -1,21 +1,10 @@
 <?php
 
-    // Récupération du paramètre NomEleve passé dans la requête ajax dans mon fichier.js
+if($_GET['ajax'] == 1){
     $id=$_GET['id_besoin'];
     $retour = true;
 
-    //$bdd = Model::$bdd;
-    $login = "root";
-    $host = "localhost";
-    $dbname = "workshop";
-    $pass = "root";
-    $dsn = "mysql:dbname=$dbname;host=$host";
-    // Connexion à la base de données    
-    try {
-        $bdd = new PDO($dsn, $login, $pass);
-    } catch (PDOException $e) {
-        echo 'Connexion échouée : ' . $e->getMessage();
-    }
+    $bdd = Model::$bdd;
 
     $req_prep = $bdd->prepare("DELETE FROM fiches_besoin WHERE id_besoin = :id");
     $req_prep->execute(array('id' => $id));
@@ -26,4 +15,6 @@
     $json = array('retour' => $retour);
 
     echo json_encode($json);
+}
+
 ?>
